refactor(refactor-type): 타입 재설정
변경된 Money 값 객체의 내부 필드 타입에 따라 파싱을 int에서 long으로 변경하였습니다.

+++ src/main/java/domain/bet/Bet.java
@@ -13,7 +13,7 @@ public class Bet {
         this.betHistory = new BetHistory(playerNames);
     }
 
-    public void bettingMoney(Name playerName, int bettingMoney) {
+    public void bettingMoney(Name playerName, long bettingMoney) {
         betHistory.bettingMoney(playerName, bettingMoney);
     }
